+++ docs/SMI.rst
@@ -1,0 +1,7 @@
+SMI plotting function
+======================
+
+This is the SMI plotting function.
+
+.. automodule:: hoggormplot.SMI_plot
+   :members:
