+++ docs/conveniencePlots.rst
@@ -1,0 +1,79 @@
+Convenience plotting functions
+==============================
+
+This is an overview of available **convenience plotting functions** in 
+HoggormPlot. They are based on the main plot function 
+``hoggorm.main_plot.plot`` and provide simplified interfaces for interactive
+use.
+
+
+Explained variance plot
+-----------------------
+
+.. note:: The two functions below, i.e. ``.explVar`` and ``.explainedVariance``
+          are two options of how to generate explained variance plots. They 
+          generate identical plots when provided the same input paramters
+          or arguments.
+
+.. automodule:: hoggormplot.conv_explainedVariancePlot
+   :members:
+
+
+Scores plot
+-----------
+
+.. automodule:: hoggormplot.conv_scoresPlot
+   :members:
+
+
+Loadings plot
+-------------
+
+.. automodule:: hoggormplot.conv_loadingsPlot
+   :members:
+
+
+Correlation loadings plot
+-------------------------
+
+.. automodule:: hoggormplot.conv_correlationLoadingsPlot
+   :members:
+
+
+Biplot
+------
+
+.. automodule:: hoggormplot.conv_biPlot
+   :members:
+
+
+Coefficients plot
+-----------------
+
+.. note:: The two functions below, i.e. ``.coeff`` and ``.coefficients`` are two
+          options of how to generate explained variance plots. They generate
+          identical plots when provided the same input paramters or arguments.
+
+.. automodule:: hoggormplot.conv_coefficientsPlot
+   :members:
+
+
+Loading weights plot
+--------------------
+
+.. automodule:: hoggormplot.conv_loadingWeightsPlot
+   :members:
+
+
+Predict plot
+------------
+
+.. automodule:: hoggormplot.conv_predictPlot
+   :members:
+
+
+
+
+
+
+
