+++ docs/mainPlot.rst
@@ -1,0 +1,11 @@
+Main plotting function
+======================
+
+This is the main plotting function for visualsation of results from models
+computed with the `Hoggorm`_ package. All convenience functions make use of 
+the main plotting function.
+
+.. _Hoggorm: http://hoggorm.readthedocs.io/en/latest/
+
+.. automodule:: hoggormplot.main_plot
+   :members:
