--- docs/plottingFuncs.rst
@@ -1,7 +1,0 @@
-Plotting functions
-==================
-
-Info on plotting functions for HoggormPlot
-
-.. automodule:: hoggormPlot.resPlotting
-   :members:
