+++ hoggormplot/__init__.py
@@ -11,5 +11,13 @@
 
 from .version import __version__
 
-from .resPlotting import (plot, scores, loadings, loadingWeights, correlationLoadings, biplot, coefficients, coeffs, explainedVariance, explVar, predict, plotSMI)
-
+from .conv_biPlot import biplot
+from .conv_coefficientsPlot import (coefficients, coeffs)
+from .conv_correlationLoadingsPlot import correlationLoadings
+from .conv_explainedVariancePlot import (explainedVariance, explVar)
+from .conv_loadingWeightsPlot import loadingWeights
+from .conv_loadingsPlot import loadings
+from .conv_predictPlot import predict
+from .conv_scoresPlot import scores
+from .main_plot import plot
+from .SMI_plot import plotSMI
